# services/web-bff/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HooshiX Web BFF Public API
  version: 2.0.0
  description: Canonical browser-facing REST contract owned by Web BFF. All examples use synthetic .test identities and non-production placeholder credentials.
servers:
  - url: /
security:
  - BrowserSession: []
paths:
...
  /api/v1/auth/oidc/google/start:
    post:
      operationId: startGoogleOidcLogin
      summary: Create a bounded Google login authorization transaction
      security: [{ BrowserSession: [] }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/auth/oidc/google/start, body: { returnTarget: /welcome } }
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/OidcStartRequest' } } }
      responses:
        '200':
          description: Provider authorization URL and bounded transaction expiry.
          content: { application/json: { schema: { $ref: '#/components/schemas/OidcStartResponse' } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/auth/oidc/google/callback:
    get:
      operationId: completeGoogleOidcCallback
      summary: Consume the single-use provider callback and rotate browser security state
      security: [{ BrowserSession: [] }]
      x-hooshix-consumer-example: { method: GET, path: '/api/v1/auth/oidc/google/callback?code=synthetic-code&state=synthetic-state' }
      parameters:
        - { name: code, in: query, required: false, schema: { type: string, minLength: 1, maxLength: 2048 } }
        - { name: state, in: query, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_-]{43}$' } }
        - { name: error, in: query, required: false, schema: { type: string, minLength: 1, maxLength: 256 } }
      responses:
        '303': { description: Session rotated and browser redirected to a validated same-origin relative path. }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/external-identities:
    get:
      operationId: getExternalIdentityStatus
      summary: Read the authenticated account's external-identity status
      security: [{ BrowserSession: [] }]
      x-hooshix-consumer-example: { method: GET, path: /api/v1/identity/external-identities }
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '200':
          description: Current Google binding status.
          content: { application/json: { schema: { $ref: '#/components/schemas/ExternalIdentityStatusResponse' } } }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/external-identities/google/start:
    post:
      operationId: startGoogleExternalIdentityLink
      summary: Create a recent-authenticated Google account-link transaction
      security: [{ BrowserSession: [] }]
      x-hooshix-consumer-example: { method: POST, path: /api/v1/identity/external-identities/google/start, body: { returnTarget: /settings/security } }
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/OidcStartRequest' } } }
      responses:
        '200':
          description: Provider authorization URL and bounded transaction expiry.
          content: { application/json: { schema: { $ref: '#/components/schemas/OidcStartResponse' } } }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
  /api/v1/identity/external-identities/google:
    delete:
      operationId: unlinkGoogleExternalIdentity
      summary: Unlink Google after recent authentication when another method remains
      security: [{ BrowserSession: [] }]
      x-hooshix-consumer-example: { method: DELETE, path: /api/v1/identity/external-identities/google }
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/CsrfToken' }]
      responses:
        '204': { description: Google identity unlinked and browser security state rotated. }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/InvalidSession' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/RegistrationRejected' }
        '503': { $ref: '#/components/responses/DependencyUnavailable' }
components:
  securitySchemes:
    BrowserSession:
      type: apiKey
      in: cookie
      name: __Host-sajtech-session
      description: HttpOnly same-origin session. Its value is never exposed in this contract or browser storage.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Caller-generated UUIDv4 idempotency identity. Replays use the same value.
      schema:
        type: string
        format: uuid
        pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Current session-bound synchronizer token returned by a reviewed BFF response; do not persist it.
      schema: { type: string, minLength: 32, maxLength: 512 }
...
  schemas:
    OidcStartRequest:
      type: object
      additionalProperties: false
      required: [returnTarget]
      properties:
        returnTarget: { type: string, minLength: 1, maxLength: 1024, pattern: '^/(?!/).*$' }
      example: { returnTarget: /welcome }
    OidcStartResponse:
      type: object
      additionalProperties: false
      required: [authorizationUrl, expiresAt]
      properties:
        authorizationUrl: { type: string, format: uri, maxLength: 4096 }
        expiresAt: { type: string, format: date-time }
      example: { authorizationUrl: 'https://accounts.google.com/o/oauth2/v2/auth?client_id=synthetic.apps.googleusercontent.com&state=synthetic', expiresAt: '2026-08-26T08:10:00Z' }
    ExternalIdentityStatusResponse:
      type: object
      additionalProperties: false
      required: [googleLinked]
      properties:
        googleLinked: { type: boolean }
      example: { googleLinked: true }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code]
      example: { type: urn:hooshix:problem:invalid-request, title: Invalid request, status: 400, code: INVALID_REQUEST, instance: /api/v1/example }
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          maxLength: 120
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{0,63}$'
        instance:
          type: string
          maxLength: 1024
  responses:
    InvalidRequest:
      description: Request validation failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InvalidSession:
      description: A supplied browser session is invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Browser-origin or session-bound request protection rejected the request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RegistrationRejected:
      description: Registration precondition was not satisfied. Response is non-enumerating.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    RateLimited:
      description: The applicable semantic security quota was exceeded.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    DependencyUnavailable:
      description: The BFF runtime or an authoritative downstream dependency is unavailable.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...